libsel4vm: pull gic dist virt into own file
This code is common between the gicv2 and giv3
